Inital Attack InterFace

+++ ArmiesHeaders/unit.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <iostream>
 using namespace std;
-
 class Game;
 
 class unit
@@ -17,14 +16,14 @@ public:
 		AM = 5,
 		AD = 6
 	};
-	unit(int id , UnitType type , int Tj  , double Health , int AttackCapacity , int AttackPower,Game*game);
+	unit(int id , UnitType type , int Tj  , double Health , int AttackCapacity , double AttackPower, Game*game);
 	virtual bool attack() = 0;
 	virtual bool getAttacked(unit* Attacker);
 	virtual bool isDead() const; // Implementes As it's the same for all
 	double getHealth() const;
-	int getPower() const;
+	double getPower() const;
 	int getID() const;
-	bool reduceHealth(int amount); // returns true if the unit died
+	bool reduceHealth(double amount); // returns true if the unit died
 	UnitType GetType() const;
 
 
@@ -40,7 +39,7 @@ protected:
 		int Td;
 		double Health;
 		int AttackCapacity;
-		int AttackPower;
+		double AttackPower;
 		Game* game;
